* Added wand alliance. Spells cast with a wand allied to someone who is not you are half as powerful. * Expelliarmus changes wand allegiance if a wand was disarmed. * Informous now has a range that changes with the level of the spell. * Informous will list the magical effects currently on a player. * Changed OllivandersPlayerListener.wandCheck return value to double, from int * Implemented Elder Wand, doubles the power of any spell if it is allied with you, if not, then it's power is equal to a wand of your type that is allied with you. This is given in the kit. * Implemented config.yml which will carry a list of worlds where the plugin is active. A empty list means that it is active for all worlds.

+++ Ollivanders/src/Spell/ACCIO.java
@@ -17,7 +17,7 @@ import org.bukkit.entity.Player;
  */
 public class ACCIO extends SpellProjectile implements Spell{
 	
-	public ACCIO(Ollivanders p, Player player, Spells name, Integer rightWand){
+	public ACCIO(Ollivanders p, Player player, Spells name, Double rightWand){
 		super(p, player, name, rightWand);
 	}
