+++ Ollivanders/src/Spell/AVADA_KEDAVRA.java
@@ -19,7 +19,7 @@ import org.bukkit.entity.Player;
 public class AVADA_KEDAVRA extends SpellProjectile implements Spell{
 
 	@SuppressWarnings("deprecation")
-	public AVADA_KEDAVRA(Ollivanders p, Player player, Spells name, Integer rightWand){
+	public AVADA_KEDAVRA(Ollivanders p, Player player, Spells name, Double rightWand){
 		super(p, player, name, rightWand);
 		moveEffectData = Material.MELON_BLOCK.getId();
 	}
