+++ Ollivanders/src/Spell/EBUBLIO.java
@@ -16,7 +16,7 @@ import org.bukkit.potion.PotionEffectType;
  */
 public class EBUBLIO extends SpellProjectile implements Spell{
 	
-	public EBUBLIO(Ollivanders p, Player player, Spells name, Integer rightWand){
+	public EBUBLIO(Ollivanders p, Player player, Spells name, Double rightWand){
 		super(p, player, name, rightWand);
 	}
